Move pkg/v1/google -> pkg/authn/google
- update pkg/authn/README
- include new modules in hack/presubmit.sh
